+++ src/Custom/VectorStores/Internal/Pagination/VectorStoreFileBatchesPageEnumerator.cs
@@ -121,7 +121,7 @@ internal partial class VectorStoreFileBatchesPageEnumerator : PageEnumerator<Vec
         request.Method = "GET";
         var uri = new ClientUriBuilder();
         uri.Reset(_endpoint);
-        uri.AppendPath("/vector_stores/", false);
+        uri.AppendPath("/v1/vector_stores/", false);
         uri.AppendPath(vectorStoreId, true);
         uri.AppendPath("/file_batches/", false);
         uri.AppendPath(batchId, true);